+++ src/search_fixed_replace.cpp
@@ -102,13 +102,13 @@ SEXP stri__replace_allfirstlast_fixed_byte(SEXP str, SEXP pattern, SEXP replacem
       deque<R_len_t_x2>::iterator iter = occurences.begin();
       for (; iter != occurences.end(); ++iter) {
          R_len_t_x2 match = *iter;
-         memcpy(curbuf, str_cur_s+jlast, match.v1-jlast);
+         memcpy(curbuf, str_cur_s+jlast, (size_t)match.v1-jlast);
          curbuf += match.v1-jlast;
          jlast = match.v2;
-         memcpy(curbuf, replacement_cur_s, replacement_cur_n);
+         memcpy(curbuf, replacement_cur_s, (size_t)replacement_cur_n);
          curbuf += replacement_cur_n;
       }
-      memcpy(curbuf, str_cur_s+jlast, str_cur_n-jlast);
+      memcpy(curbuf, str_cur_s+jlast, (size_t)str_cur_n-jlast);
       SET_STRING_ELT(ret, i, Rf_mkCharLenCE(buf.data(), buf_need, CE_UTF8));
    }
